+++ src/rockverse/configure/latex_strings.py
@@ -1,0 +1,3 @@
+LATEX_STRINGS = {
+    'um': r'$\mu$m',
+}